feat: modify modal component, modify ts file, modify package.json
modify modal component, modify ts file, modify package.json
