+++ packages/core/src/components/Input/index.tsx
@@ -1,9 +1,38 @@
-import { Input as AntdInput } from 'antd'
+import { Input as AntdInput, InputProps } from 'antd'
 import CircleClose from 'assets/circle-close.svg?react'
 import EyeTwoTone from 'assets/eye.svg?react'
 import EyeInvisibleOutlined from 'assets/eye-invisible.svg?react'
 import useStyles from './style'
-import { IInputProps, InputPasswordProps, InputTextAreaProps } from './Input'
+import type { TextAreaProps, PasswordProps } from 'antd/lib/input'
+import { ReactNode } from 'react'
+
+export type InputSizeType = 'small' | 'middle'
+
+export interface IInputProps
+  extends Omit<InputProps, 'size' | 'onClear' | 'allowClear'> {
+  size?: InputSizeType
+  onClear?: React.MouseEventHandler<HTMLElement>
+  count?: CountConfig
+}
+
+export type InputPasswordProps = Omit<PasswordProps, 'size' | 'allowClear'> &
+  IInputProps
+
+export type InputTextAreaProps = Omit<TextAreaProps, 'size' | 'allowClear'> &
+  IInputProps
+
+export interface CountConfig {
+  max?: number
+  strategy?: (value: string) => number
+  show?:
+    | boolean
+    | ((args: {
+        value: string
+        count: number
+        maxLength?: number
+      }) => ReactNode)
+  exceedFormatter?: (value: string, config: { max: number }) => string
+}
 
 const getClearIcon = (
   onClear?: React.MouseEventHandler<HTMLElement> | undefined
