--- packages/core/src/components/Modal/Modal.d.ts
@@ -1,5 +1,0 @@
-import { ModalProps } from 'antd'
-
-export interface IModalProps extends Omit<ModalProps, 'wrapClassName'> {
-  wrapClassName?: string
-}
