+++ packages/core/src/components/Modal/index.tsx
@@ -1,6 +1,12 @@
-import { Modal as AntdModal } from 'antd'
-import { IModalProps } from './Modal'
+import { Modal as AntdModal, ModalProps } from 'antd'
+import Close from 'assets/close.svg?react'
 import useStyles from './style'
+import { useTheme } from 'antd-style'
+
+export interface IModalProps
+  extends Omit<ModalProps, 'wrapClassName' | 'closeIcon'> {
+  wrapClassName?: string
+}
 
 function Modal({
   wrapClassName,
@@ -9,11 +15,18 @@ function Modal({
   ...props
 }: IModalProps) {
   const { styles, cx } = useStyles()
+  const token = useTheme()
+  console.log(token)
   return (
     <AntdModal
       {...props}
       width={width}
       wrapClassName={cx(styles.modalWrap, wrapClassName)}
+      closeIcon={
+        <div className={styles.modalWrap}>
+          <Close color={token?.colorTextBase} width="16" height="16" />
+        </div>
+      }
     >
       {children}
     </AntdModal>
