--- packages/core/src/components/Tabs/Tabs.d.ts
@@ -1,9 +1,0 @@
-import { TabsProps } from 'antd'
-
-export type TabsSizeType = 'middle' | 'small'
-
-export interface ITabsProps
-  extends Omit<TabsProps, 'size' | 'tabBarGutter' | 'rootClassName'> {
-  size?: TabsSizeType
-  rootClassName?: string
-}
