+++ packages/core/src/components/ToolTip/index.tsx
@@ -1,9 +1,9 @@
 import { Tooltip, TooltipProps as antdTooltipProps } from 'antd'
 
-export type TooltipProps = antdTooltipProps
+export type ITooltipProps = antdTooltipProps
 import useStyles from './style'
 
-export default function ToolTip(props: TooltipProps) {
+export default function ToolTip(props: ITooltipProps) {
   const { styles, cx } = useStyles()
   const { children } = props
   return (
