+++ packages/core/src/components/index.tsx
@@ -1,10 +1,29 @@
 export { default as Button } from './Button'
+export type { IButtonProps } from './Button'
 export { default as Collapse } from './Collapse'
+export type { ICollapseProps } from './Collapse'
 export { default as Dropdown } from './Dropdown'
+export type { IDropdownProps } from './Dropdown'
 export { default as Input } from './Input'
+export type {
+  IInputProps,
+  InputPasswordProps,
+  InputTextAreaProps
+} from './Input'
 export { default as Modal } from './Modal'
+export type { IModalProps } from './Modal'
 export { default as Pagination } from './Pagination'
+export type { IEpPaginationProps } from './Pagination'
 export { default as Search } from './Search'
+export type { ISearchProps } from './Search'
 export { default as Table } from './Table'
+export type { ITableProps } from './Table'
 export { default as Tabs } from './Tabs'
+export type { ITabsProps } from './Tabs'
+export { default as ToolTip } from './ToolTip'
+export type { ITooltipProps } from './ToolTip'
 export { default as Typography } from './Typography'
+export type { ITextProps } from './Typography/Text'
+export type { IAelfdTitleProps } from './Typography/Title'
+export { default as AELFDProvider } from './provider'
+export type { IAelfdThemeProviderProps } from './provider'
