+++ packages/core/src/examples/App.tsx
@@ -1,20 +1,20 @@
 import Pagination from 'components/Pagination'
 import { SearchOutlined, AndroidOutlined } from '@ant-design/icons'
-import Button from './Button'
-import Input from './Input'
-import Dropdown from './Dropdown'
-import Typography, { FontWeightType } from './Typography'
-import ConfigProvider from '../provider'
+import Button from 'components/Button'
+import Input from 'components/Input'
+import Dropdown from 'components/Dropdown'
+import Typography, { FontWeightType } from 'components/Typography'
+import ConfigProvider from '../components/provider'
 import { ThemeAppearance } from 'antd-style'
 import { Segmented } from 'antd'
 import { useState } from 'react'
-import Collapse, { ICollapseProps } from './Collapse'
-import Search from './Search'
-import Tabs from './Tabs'
+import Collapse, { ICollapseProps } from 'components/Collapse'
+import Search from 'components/Search'
+import Tabs from 'components/Tabs'
 import DownArrow from 'assets/downArrow.svg?react'
 import TableExample from 'examples/tableExample'
 import ModalExample from 'examples/modalExample'
-import ToolTipExample from '../examples/tooltipExample'
+import ToolTipExample from 'examples/tooltipExample'
 
 const items = [
   {
