+++ packages/core/src/examples/index.tsx
@@ -1,6 +1,6 @@
 import { createRoot } from 'react-dom/client'
 import 'tailwindcss/tailwind.css'
-import App from 'components/App'
+import App from 'examples/App'
 
 const container = document.getElementById('root') as HTMLDivElement
 const root = createRoot(container)
